simplify with bash strict mode

+++ scripts/installs/install-firefox.sh
@@ -2,6 +2,7 @@
 
 set -euo pipefail
 
+command -v firefox && exit 0 || true
 
 # 'https://download.mozilla.org/?product=firefox-latest&os=linux64&lang=en-US'
 # URL="https://download.mozilla.org/?product=firefox-52.1.1esr-SSL&os=linux64&lang=en-US"
@@ -9,27 +10,22 @@ URL="https://download.mozilla.org/?product=firefox-45.9.0esr-SSL&os=linux64&lang
 cd /tmp/
 FFTAR="firefox.tar.bz2"
 FF=$(basename "${FFTAR}" .tar.bz2)
+test -n "${FF}"
 
-if test -f ${FFTAR}; then
-    rm -f ${FFTAR} || exit ${LINENO}
-    true
-fi
+test -f ${FFTAR} && rm -f ${FFTAR}
 
-wget "${URL}" -O "${FFTAR}" || exit ${LINENO}
+wget "${URL}" -O "${FFTAR}"
 
 INSTALL_DIR=/usr/local/${FF}
 
-if test -d ${INSTALL_DIR}; then
-    test -n "${FF}" || exit ${LINENO}
-    sudo rm -rf /usr/local/${FF}
-fi
+test -d ${INSTALL_DIR} && sudo rm -rf /usr/local/${FF}
 
-sudo tar -C /usr/local -axvf "${FFTAR}" || exit ${LINENO}
-cd  || exit ${LINENO}
+sudo tar -C /usr/local -axvf "${FFTAR}"
+cd
 
 
 EXE="${INSTALL_DIR}/firefox-bin"
-test -f ${EXE} || exit ${LINENO}
+test -f ${EXE}
 
 sudo update-alternatives --install /usr/local/bin/firefox firefox "${EXE}" 20
 sudo update-alternatives --set firefox "${EXE}"
